refactor(vendingmachine): 패키지 분리
controller와 domain 패키지를 추가하여 클래스를 이동시켰습니다.
