+++ src/main/java/vendingmachine/domain/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.domain;
 
 import java.util.Arrays;
 import java.util.List;
