+++ src/main/java/vendingmachine/domain/Item.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.domain;
 
 public class Item {
 	private final int price;
